Export FastAPI router
